ai-agents-crew-ai-apps/sales_pitch_assistant/backend/requirements.txt
crewai>=0.28.0
crewai-tools>=0.8.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-groq>=0.0.1
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
